add projection manager
