+++ src/Container/PostgresEventStoreFactory.php
@@ -34,12 +34,8 @@ final class PostgresEventStoreFactory extends AbstractEventStoreFactory
                 TransactionalActionEventEmitterEventStore::EVENT_HAS_STREAM,
                 TransactionalActionEventEmitterEventStore::EVENT_FETCH_STREAM_METADATA,
                 TransactionalActionEventEmitterEventStore::EVENT_UPDATE_STREAM_METADATA,
-                TransactionalActionEventEmitterEventStore::EVENT_DELETE_PROJECTION,
-                TransactionalActionEventEmitterEventStore::EVENT_RESET_PROJECTION,
-                TransactionalActionEventEmitterEventStore::EVENT_STOP_PROJECTION,
                 TransactionalActionEventEmitterEventStore::EVENT_FETCH_STREAM_NAMES,
                 TransactionalActionEventEmitterEventStore::EVENT_FETCH_CATEGORY_NAMES,
-                TransactionalActionEventEmitterEventStore::EVENT_FETCH_PROJECTION_NAMES,
                 TransactionalActionEventEmitterEventStore::EVENT_BEGIN_TRANSACTION,
                 TransactionalActionEventEmitterEventStore::EVENT_COMMIT,
                 TransactionalActionEventEmitterEventStore::EVENT_ROLLBACK,
@@ -88,7 +84,6 @@ final class PostgresEventStoreFactory extends AbstractEventStoreFactory
             ],
             'load_batch_size' => 1000,
             'event_streams_table' => 'event_streams',
-            'projections_table' => 'projections',
             'message_factory' => FQCNMessageFactory::class,
             'wrap_action_event_emitter' => true,
             'metadata_enrichers' => [],
